refactor(agent): extract fallback policy into auxiliary_fallback_policy (auxiliary_client.py R3-S1, epic #78647)
